+++ src/plugins/chess_results/chess_results_background_uploader.py
@@ -146,7 +146,7 @@ class ChessResultsBackgroundUploader:
         event_uniq_id: str,
         tournament_id: int,
         force: bool,
-    ) -> None:
+    ) -> ChessResultsUploadResult | None:
         """Upload a tournament to Chess-Results.com."""
 
         # Set the locale (called in a new thread)
@@ -156,18 +156,18 @@ class ChessResultsBackgroundUploader:
         loader = EventLoader()
         if event_uniq_id not in loader.event_uniq_ids:
             # The event has been deleted
-            return
+            return None
         event = loader.load_event(event_uniq_id)
 
         tournament = event.tournaments_by_id.get(tournament_id, None)
         if not tournament:
             # The tournament has been deleted
-            return
+            return None
 
         current_result = cls.get_updated_tournament_upload_result(tournament)
         if current_result.status == ChessResultsUploadStatus.SETTINGS_ERROR:
             # Skip this tournament if we now have a SETTINGS_ERROR
-            return
+            return current_result
 
         result_id = cls.result_id(tournament.event.uniq_id, tournament.id)
         if not force and not ChessResultsUtils.resolve_auto_upload(tournament):
@@ -176,7 +176,7 @@ class ChessResultsBackgroundUploader:
                 ChessResultsUploadStatus.CHANGED,
                 _('Modified since last upload'),
             )
-            return
+            return cls.upload_status_messages[result_id]
 
         if not NetworkMonitor.connected():
             # The network is offline, we can't upload
@@ -185,7 +185,7 @@ class ChessResultsBackgroundUploader:
                 _('Modified, but no internet connection'),
             )
             cls.publish_upload_event()
-            return
+            return cls.upload_status_messages[result_id]
 
         cls.upload_status_messages[result_id] = ChessResultsUploadResult(
             ChessResultsUploadStatus.IN_PROGRESS,
@@ -221,6 +221,8 @@ class ChessResultsBackgroundUploader:
             )
         finally:
             cls.publish_upload_event()
+
+        return cls.upload_status_messages[result_id]
 
     @classmethod
     def upload_event(cls, admin_event: Event) -> None: