+++ src/web/controllers/admin/prize_admin_controller.py
@@ -20,7 +20,7 @@ from data.print_documents.documents import (
     PrizeListPrintDocument,
 )
 from data.prize.managers import PrizeSharingManager
-from data.criteria.managers import PlayerFilterManager, PlayerFilterOptionManager
+from data.criteria.managers import PrizePlayerFilterManager, PlayerFilterOptionManager
 from data.criteria.player_filters import PlayerFilter
 from data.prize.prize import Prize
 from data.prize.prize_category import PrizeCategory
@@ -808,7 +808,7 @@ class PrizeAdminController(BaseEventAdminController):
         field = 'type'
         player_filter_id = data.get(field, '')
         try:
-            PlayerFilterManager(event).get_type(player_filter_id)
+            PrizePlayerFilterManager(event).get_type(player_filter_id)
         except KeyError:
             errors[field] = _('Please select a type of criterion.')
             return errors
@@ -821,7 +821,7 @@ class PrizeAdminController(BaseEventAdminController):
 
     @staticmethod
     def player_filter_from_data(event: Event, data: dict[str, str]) -> PlayerFilter:
-        player_filter_type = PlayerFilterManager(event).get_type(data['type'])
+        player_filter_type = PrizePlayerFilterManager(event).get_type(data['type'])
         options = []
         for option in player_filter_type().default_options():
             value = WebContext.form_data_to_value(data, option.id, option.type)
@@ -844,13 +844,13 @@ class PrizeAdminController(BaseEventAdminController):
             'modal': 'prize_criterion_form',
             'action': action,
             'player_filter_select_options': {'': '-'}
-            | PlayerFilterManager(event).options(),
+            | PrizePlayerFilterManager(event).options(),
             'player_filter_options': PlayerFilterOptionManager(event).objects(),
             'containers_by_type': {
                 player_filter.id: [
                     option.container_id for option in player_filter.default_options()
                 ]
-                for player_filter in PlayerFilterManager(event).objects()
+                for player_filter in PrizePlayerFilterManager(event).objects()
             }
             | {'': []},
             'add_other_active': (