+++ RELEASE_NOTES.md
@@ -1,14 +1,24 @@
 # French
 English version below
 
+## FFE
+
+- Ajout d’une option pour mettre en ligne via le menu d’action de la fiche tournoi (3.3.0)
+
 ## _Chess-Results_
 
 - Ajout de QR-codes vers les tournois en ligne (3.3.0)
+- Ajout d’une option pour mettre en ligne via le menu d’action de la fiche tournoi (3.3.0)
 
 ---
 
 # English
 
+## FFE
+
+- Added option to upload via the tournament card action menu (3.3.0)
+
 ## _Chess-Results_
 
 - Added QR-codes to the tournaments online (3.3.0)
+- Added option to upload via the tournament card action menu (3.3.0)